plans: split local_scans.py per category (#56)
`local_scans.py` had grown to 1455 lines covering moves, base scans,
grid scans, hklpy2-specific scans, and the merge of `multi_issues` +
`hklpy2` made it worse.  Per the issue, split it by category:

- `move_plans.py`  — `mv`, `mvr`, `abs_set` (127 lines)
- `base_scans.py`  — `count`, `ascan`, `lup`, `qxscan` (495 lines)
- `grid_scans.py`  — `grid_scan`, `rel_grid_scan` (275 lines)
- `hkl_scans.py`   — `th2th`, `hklscan`, `hscan`, `kscan`, `lscan`,
                     `psiscan` (649 lines)

`local_scans.py` is now a 51-line backward-compat shim that re-exports
every public symbol so existing
`from id4_common.plans.local_scans import ...` callers (downstream
scripts, beamline macros) keep working.

Internal callers updated to import directly from the new modules:

- `plans/__init__.py`         (the `from id4_common.plans import *` path)
- `plans/peak_position.py`    (was `from .local_scans import mv`)
- `plans/center_maximum.py`   (same)
- `plans/workflow_plan.py`    (same)
- `plans/flyscan_demo.py`     (same)

Dependency graph (no cycles):

  move_plans  ←  base_scans  ←  hkl_scans
                   grid_scans (independent)

All existing private helpers (`_PSI_MODES`, `_single_axis_scan`)
travelled with the public functions that use them.  No behaviour
changes — function bodies, signatures, and docstrings are byte-identical
to their previous in-`local_scans.py` versions.

Verified: `pre-commit run --all-files` clean, `pytest tests/` 18/18
pass, every public scan importable from both the new module *and* the
shim (and `is`-identical between the two).
