+++ src/id4_common/plans/__init__.py
@@ -2,23 +2,23 @@
 local, custom Bluesky plans (scans) and other functions
 """
 
+from .base_scans import ascan  # noqa: F401
+from .base_scans import count  # noqa: F401
+from .base_scans import lup  # noqa: F401
+from .base_scans import qxscan  # noqa: F401
 from .center_maximum import cen  # noqa: F401
 from .center_maximum import maxi  # noqa: F401
-from .local_scans import abs_set  # noqa: F401
-from .local_scans import ascan  # noqa: F401
-from .local_scans import count  # noqa: F401
-from .local_scans import grid_scan  # noqa: F401
-from .local_scans import hklscan  # noqa: F401
-from .local_scans import hscan  # noqa: F401
-from .local_scans import kscan  # noqa: F401
-from .local_scans import lscan  # noqa: F401
-from .local_scans import lup  # noqa: F401
-from .local_scans import mv  # noqa: F401
-from .local_scans import mvr  # noqa: F401
-from .local_scans import psiscan  # noqa: F401
-from .local_scans import qxscan  # noqa: F401
-from .local_scans import rel_grid_scan  # noqa: F401
-from .local_scans import th2th  # noqa: F401
+from .grid_scans import grid_scan  # noqa: F401
+from .grid_scans import rel_grid_scan  # noqa: F401
+from .hkl_scans import hklscan  # noqa: F401
+from .hkl_scans import hscan  # noqa: F401
+from .hkl_scans import kscan  # noqa: F401
+from .hkl_scans import lscan  # noqa: F401
+from .hkl_scans import psiscan  # noqa: F401
+from .hkl_scans import th2th  # noqa: F401
+from .move_plans import abs_set  # noqa: F401
+from .move_plans import mv  # noqa: F401
+from .move_plans import mvr  # noqa: F401
 from .peak_position import peak  # noqa: F401
 from .peak_position import peak_pos  # noqa: F401
 from .peak_position import pmax  # noqa: F401
