+++ src/id4_common/plans/move_plans.py
@@ -1,0 +1,127 @@
+"""
+Local move plans: ``mv``, ``mvr``, ``abs_set``.
+
+Thin wrappers around the corresponding bluesky plan stubs that also stage
+the 9-Tesla magnet (``magnet911``) when its field is one of the targets.
+"""
+
+__all__ = [
+    "mv",
+    "mvr",
+    "abs_set",
+]
+
+from bluesky.plan_stubs import abs_set as bps_abs_set
+from bluesky.plan_stubs import mv as bps_mv
+from bluesky.preprocessors import relative_set_decorator
+from toolz import partition
+
+from ._local_scan_utils import _check_magnet911
+from .local_preprocessors import stage_magnet911_decorator
+
+
+def mv(*args, **kwargs):
+    """
+    Move one or more devices to a setpoint, and wait for all to complete.
+
+    This is a local version of `bluesky.plan_stubs.mv`. If more than one device
+    is specifed, the movements are done in parallel.
+
+    Parameters
+    ----------
+    args :
+        device1, value1, device2, value2, ...
+    kwargs :
+        passed to bluesky.plan_stubs.mv
+
+    Yields
+    ------
+    msg : Msg
+
+    See Also
+    --------
+    :func:`bluesky.plan_stubs.mv`
+    """
+
+    magnet_option = _check_magnet911(args)
+
+    @stage_magnet911_decorator(magnet_option)
+    def _inner_mv():
+        yield from bps_mv(*args, **kwargs)
+
+    return (yield from _inner_mv())
+
+
+def mvr(*args, **kwargs):
+    """
+    Move one or more devices to a relative setpoint. Wait for all to complete.
+
+    If more than one device is specified, the movements are done in parallel.
+
+    This is a local version of `bluesky.plan_stubs.mvr`.
+
+    Parameters
+    ----------
+    args :
+        device1, value1, device2, value2, ...
+    kwargs :
+        passed to bluesky.plan_stub.mvr
+
+    Yields
+    ------
+    msg : Msg
+
+    See Also
+    --------
+    :func:`bluesky.plan_stubs.rel_set`
+    :func:`bluesky.plan_stubs.mv`
+    """
+    objs = []
+    for obj, _ in partition(2, args):
+        objs.append(obj)
+
+    @relative_set_decorator(objs)
+    def _inner_mvr():
+        return (yield from mv(*args, **kwargs))
+
+    return (yield from _inner_mvr())
+
+
+def abs_set(*args, **kwargs):
+    """
+    Set a value. Optionally, wait for it to complete before continuing.
+
+    This is a local version of `bluesky.plan_stubs.abs_set`. If more than one
+    device is specifed, the movements are done in parallel.
+
+    Parameters
+    ----------
+    obj : Device
+    group : string (or any hashable object), optional
+        identifier used by 'wait'
+    wait : boolean, optional
+        If True, wait for completion before processing any more messages.
+        False by default.
+    args :
+        passed to obj.set()
+    kwargs :
+        passed to obj.set()
+
+    Yields
+    ------
+    msg : Msg
+
+    See Also
+    --------
+    :func:`bluesky.plan_stubs.rel_set`
+    :func:`bluesky.plan_stubs.wait`
+    :func:`bluesky.plan_stubs.mv`
+    """
+
+    magnet_option = _check_magnet911(args)
+
+    @stage_magnet911_decorator(magnet_option, persistent=False)
+    def _inner_abs_set():
+        yield from bps_abs_set(*args, **kwargs)
+
+    return (yield from _inner_abs_set())
